+++ CHANGELOG.md
@@ -1,3 +1,58 @@
+## VIKING [0.1.20]
+
+### Добавлено
+- Commands.js:
+    - Команда help()
+- CreepManager.js:
+    - Функция подсчета живых существ
+    - Функция запуска кода для каждой роли
+    - Новые роли
+- Файл Visual.js
+
+### Изменено
+- CreepManager.js:
+    - Добавление ролей в память перенесено из CreepManager.js в SetMemory.js
+
+### Исправлено
+- Commands.js:
+    - Команда ClearMemory()
+    - Команда ClearFlags()
+- CreepManager.js:
+    - Установка количества необходимых крипов для комнат
+- main.js:
+    - Запуск основных функций
+- RoomStats.js:
+    - Некорректное отображение некоторых параметров
+- Autobuilder.js:
+    - Незначительные исправление
+---
+
+### Added
+- Commands.js:
+    - help() command
+- CreepManager.js:
+    - Function of calculate live creeps
+    - Function of start code for all roles
+    - New roles
+- File Visual.js
+
+### Changed
+- CreepManager.js:
+    - Adding roles in memory moved from CreepManager.js in SetMemory.js
+
+### Fixed
+- Commands.js:
+    - Command ClearMemory()
+    - Command ClearFlags()
+- CreepManager.js:
+    - Setting amount need creeps for rooms
+- main.js:
+    - Starting major functions
+- RoomStats.js:
+    - Uncorrected output some parameters
+- Autobuilder.js:
+    - Not important
+
 ## VIKING [0.1.3]
 
 ### Добавлено